Fix going back more than 1 view not working
Fix navigation history not working when going back more than 1 view. If
I opened the following screen views: A -> B -> C, the current view was C
and the previous view was B but when going back to B, the previous view
was now C, this created a loop where we could never go back to A and
only B -> C -> B -> C...
This has been fixed by allowing multiple screen views to be added to the
navigation history instead of having only one (which is even better
because the screen and modal views history now work the same way). The
navigation history has a reference to all the previously opened
views, and is only cleared when the user goes back to the main menu.