Add Star.uss to handle the styles of the stars
Add the USS file Star.uss to handle the styles of the stars in all
UI. In any UXML file, the star can be now be shown as inactive (grey) or
active (gold) by adding/removing the classes from Star.uss. This will allow
for easier control and modification of the star styles.