Refactor the code to replace Task with UniTask
All default C# Task have been replace with UniTask. This has been done
because C# Tasks are multi-threaded and won't work with the WebGL
version of the game (as most browsers don't support it yet). UniTask
runs on the Unity's player loop and only works on the main thread
(except for UniTask.Run()). Therefore, This library allows us to use
await/async for all targeted platforms (instead of having to use Unity's
Coroutines).