Switch from JsonUtility to Newtonsoft.Json
JsonUtility cannot serialize/deserialize properties, only fields. This
is a problem for the new obstacles system. Newtonsoft.Json being a more
complete library, will be better suited for this task and the future.
This change has not been implemented everywhere yet, file data
persistence and level selection deserialization still use JsonUtility.
For now, it will remain this way as using both libraries for different
tasks is not a problem, but it might be changed in the future.