Separate the obstaclesContainer in 2 containers
Separate the obstaclesContainer in 2 containers, one for the absolutely
positionnable obstacles (eg: floor hole) and one for the relatively
positionnable obstacles (eg: spikes).
This has been done because relatively positionnable obstacles depend on
the absolutely positionnable obstacles to be placed. But if we only use
1 container, a reliatively positionnable obstacle can be treated before
the obstacle it is placed on top of has been saved, in which case it
would not be found.