Refactor obstacle rendering when loading levels
Instead of rendering the obstacle in the level loaders (Maze.cs or
LoadMazeFromFile.cs), the obstacle class now implements a Render method
so that the rendering of each obstacle is done in the obstacle class
itself.
This not only makes the code a lot more readable and cleaner in the
level loaders but it also makes it easier to maintain as the rendering
code was duplicatd in each loader, now both can simply call the Render
method.