Add custom converter for deserializing Obstacle
Add custom converter to instantiate the right Obstacle subclass when
deserializing from JSON.
This is needed because the obstacles are saved in the 'obstacles' array
which contains a list of Obstacle objects. But when deserializing, we
don't actually know which obstacle it is, so we need to use a custom
converter to instantiate the right subclass based on the objectType
saved in the obstacle.