Add method to instantiate an obstacle mesh/prefab
Add a static method to instantiate an obstacle's mesh or prefab when
rendering an obstacle when loading a level.
This method is needed because the Instantiate() method used for
instantiating meshes and prefabs is only available for classes that
inherit from MonoBehaviour, and the Obstacle class does not (and can't
otherwise new obstacles could not be created the way it is now as
MonoBehaviour scripts can't be instantiated with a constructor).