Fix loading level more than once not working
This commit fixes an issue where if you loaded a level, then went back
to the main menu and loaded a level again, an exception would be thrown.
This was because the Dictionary containing all obstacles GameObjects
references was not being cleared when loading a new level, and so,
GameObjects that had been destroyed were still trying be accessed.