Fix maze not centered not rotating properly
Mazes that are not centered around the origin (0, 0, 0) were being rotated around the origin instead of their center.
To fix this, I first tried to rotate the maze using RotateAround() (see 'Maze rotation' card on Trello), but it didn't feel as smooth as the previous method.
The fix that has been implemented is to move the maze to the origin when saving it to a file.