Fix issues with pause and skip views
- Fix going back in navigation history when the pause or skip view were
opened wouldn't not allow the player to resume the level. This was
because the level state would not update when going back in the
navigation history instead of closing the view using the button
- Fix pressing the close button on the pause or skip view would start
the level if it was waiting to start. This was because pressing the
close button would resume the level even if it had not started yet