Fix hide called multiples times on progress end
Once the second chance or level failed radial progress ends, the hide
method is called and the view is hidden, but the while loop still has
time to run multiple times (~70) before the view is actually disabled.
This caused the hide function to be called multiple times. This has been
fixed by simply breaking out of the loop after hiding the view.