Fix joystick invisible after unpausing/unskipping
When closing the pause or skipping view, although the joystick was not
visible, it was still enabled and could be interacted with, but the
level would only be resumed after the closing animation completed (about
half a second later). This means that if the player pressed the joystick
in the meantime, the joystick would stay invisible until the player
released it and pressed it again. To fix this issue, the level is now
resumed before the closing animation