Refactor LevelDataManager to work with all modes
This commit refactors the LevelDataManager to be more generic and work
for all modes. Indeed, the LevelDataManager was only working for the
default levels. Now, it is an abstract class that implements logic that
can be reused by most modes (primarly daily levels and events). Each
mode then inherits from this class and implements specific logic if
needed