Add local remote config file for testing
This commit adds the option to laod the remote config from a local file
instead of fetching from the server. This is useful for testing without
having to wait for the server to respond everytime.
The remote config source can be changed easily thanks to a serialized
field in GameManager