Add daily levels loading logic and UI
- Add daily levels times saving/loading
- Add the DailyLevelsDataManager and its logic
- Add daily levels specific logic for DailyLevelsLevelManager
- Implement daily levels UI to display to unlock the levels and display
the time and stars earnt...
- Add daily levels events to update the UI...
- Reset daily levels times everyday
- Refactor LevelManager to reuse the same instance for the level manager
of each mode. This allow to easily access a level manager for any mode
without having to switch the mode. For example, we can easily get the
number of stars earnt for a daily level without having to
change the mode (while before, we would have had to call SwitchMode(),
get the number of stars and call SwitchMode() again)
- Fix DefaultLevelSelectionTemplate not having its USS file linked