Refactor skins material loading
This commit moved the process of loading the skin material (from the
addressables) from the SkinView to the SkinManager. This has been done
because the SkinManager sometimes needs to access the materials (eg: to
equip the skin when starting the game). Since it seemed weird to have
the controller query the view to get the material, I moved the process
to the SkinManager.
This way, when the game loads and the skins list is loaded by the
manager, it can load and cache the skin it needs. Then, when the view
needs a material, it can query the SkinManager if the material is cached
or ask it to load it otherwise and pass a callback to get the result
once the material is loaded
Furthermore, the material is now also applied to the ball when the skin
is equipped